requirements.txt
pyodbc>=4.0.0
picamera2>=0.3.0
RPi.GPIO>=0.7.0
